update adding styling to the list item functionality
