+++ src/displayScore.js
@@ -1,15 +1,11 @@
-const board = document.querySelector('.leaderboard')
-  
+import appendScore from './appendScore'
+import addScoreStyle from './addScoreStyle'
+ 
 async function displayScores() {
   let response = await fetch('https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/jXwsUKYAzNn8VCQ6q31T/scores')
   let data = await response.json()
-  console.log(data.result)
   data.result.forEach((element) => {
-    let score = document.createElement('li')
-    score.innerHTML = `
-      ${element.user}: ${element.score}
-    `
-    board.appendChild(score)
+    appendScore(element.user, element.score)
   });
 }
